fix(types): cast to int8 to fix CI
this fix is very likely wrong, as the types are no longer encoded as
int32_t's. However, a rewrite of the type class is planned anyway,
once I understand the type system fully

+++ include/watever/type.h
@@ -153,7 +153,8 @@ public:
   //   ...
   //   (block (type $T) ...)
   //
-  bool IsIndex() const { return static_cast<int32_t>(enum_) >= 0; }
+  // TODO very likely wrong
+  bool IsIndex() const { return static_cast<int8_t>(enum_) >= 0; }
 
   Index GetIndex() const {
     assert(IsIndex());
